script: automate graphics card detection
This commit is designed to automate the graphics card detection.
As of today, there is support for NVIDIA, AMD, and Intel Graphics cards.
The code uses "lspci" and "grep" commands to find the graphics card and get its brand name, and it installs the necessary drivers and support software.
If it does not find a graphics card that matches the script, it will just skip the support software installation and let the user install it.

Signed-off-by: ninetyninebytes <74156492+ninetyninebytes@users.noreply.github.com>