cosmetic: change indents to one format
this update changes all indents to one format. the new format is using
tabs, and is unified across the entire script. before, it used to have
different indents. for example, parts of the script were 2 spaces, and
some parts of the script were 4 spaces. we've decided to just use
regular tab indents instead to make it look cleaner.

Signed-off-by: ninetyninebytes <74156492+ninetyninebytes@users.noreply.github.com>