feature: detect non root user accounts
this commit is made to stop the installation if it detects a non root
user account. this is implemented as a safety measure so that users don't
get a partially installed system, which can be dangerous and unstable.

this safety measure also makes it to prevent the script from failing if
it runs as a non root user. the script will run if it is run as a normal
user, however most of the commands require superuser access, so the
script will keep running but all of the commands will fail. the safety
measure prevents this by simply refusing to run if it detects a UID
which is not equal to 0 (the superuser UID)

Signed-off-by: ninetyninebytes <74156492+ninetyninebytes@users.noreply.github.com>