feature: change os-release to correct values
this commit is designed to change the valeus in os-release to the proper
values. by default, there is a symlink from /usr/lib/os-release to
/etc/os-release. this commit makes it so that the old files are deleted,
and the proper ones are built with the right values, and the symlink is
restored.

this is a pretty big change, because this is basically a branding
update. this will be visible to the user, so it's very important

Signed-off-by: ninetyninebytes <74156492+ninetyninebytes@users.noreply.github.com>