feature: add yay support to the script
this commit brings yay support to the script. (yay!) to get this to
work, i had to do some very complicated things.

firstly, i implemented some code that gets the username for uid 1000.
then, i used sudo with the -i and -u options combined with the
"uidtousername" variable to run all of the yay commands as the normal
user (uid 1000). this should successfully install yay as the default
user, and it should work fine.

Signed-off-by: ninetyninebytes <74156492+ninetyninebytes@users.noreply.github.com>