bugfix: potential fix for script not launching
this commit might be a fix for the script not launching. when you run
the script, it is stopped with an error on line 206. i've searched for a
solution, and this was the one i was able to come up with.

i'm making this commit here because i can't get the files over to my
virtual machine for testing. maybe it's time i get a testing machine
soon..

Signed-off-by: ninetyninebytes <74156492+ninetyninebytes@users.noreply.github.com>