feature: add yay to the script
this commit adds a bit of code that lets the user decide if they want to
install yay or not. the way this works is that once the installation is
over, it asks the user if they would like to install yay or not. if they
choose to install it, the script will install yay.

because yay requires the user to be present during the installtion, i've
added a note that lets the user know that they need to be there during
the installation so the can press Y on the prompts.

Signed-off-by: ninetyninebytes <74156492+ninetyninebytes@users.noreply.github.com>