removed testing branch notice
most likely we will be moving away from the old system, and moving to a
new branch system. currently, we're using the system where we do all of
the updates on a testing branch, and we merge to main.

while this is good, its not that efficient. so at the moment we should
consider changing our branch system so that more developers can work
better.

Signed-off-by: ninetyninebytes <74156492+ninetyninebytes@users.noreply.github.com>