installer: clarify non root error message
this commit clears up confusion in the installer. previously, the
installer would give you a warning saying "to continue this
installation, please run as root." while this is a good error message,
it can be confusing the people who are beginners and don't know what
root is.

the commit brings a newer message. the message now guides people in
the right direction on how to handle the issue. the message explains
that running as a non root user could cause a partial and unstable
installation. it also guides users to looking on the official Caveman
Linux page to see for solutions and post bugs.

Signed-off-by: ninetyninebytes <74156492+ninetyninebytes@users.noreply.github.com>