Add Linter to Codebase
This is a proposal to add a code linter for the following branches:

- testing
- release-candidate
- installer-v2

This linter works across multiple programming languages (including bash), and will help keep the code clean. This change would best work if it's added to the main branch first, and then the other branches sync their copies.