installer: move confirmation screen to the end
This commit moves the confirm screen to the end of the sequence. The
user will be presented with the summary, then they must enter the
phrase. If they fail the verification, it won't exit the installer
because then they would lose their progress. Instead, it returns back to
the summary screen. The user is allowed as many tries as they need,
until they enter the correct phrase.

Signed-off-by: spacefault <74156492+spacefault@users.noreply.github.com>